chore(dal): factor out conflict and update detection
